new java files for new schema
git-svn-id: https://oscars.es.net/repos/oscars/trunk@10008 c75a6b91-02ef-0310-b74a-e20389c9c41d

+++ nsibridge/src/main/java/org/ogf/nsi/schema/connectionTypes/ComponentPathType.java
@@ -1,7 +1,6 @@
 
 package org.ogf.nsi.schema.connectionTypes;
 
-import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
@@ -44,13 +43,13 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="directionality" type="{http://schemas.ogf.org/nsi/2011/07/connection/types}DirectionalityType"/>
- *         &lt;element name="sourceSTP" type="{http://schemas.ogf.org/nsi/2011/07/connection/types}ServiceTerminationPointType"/>
- *         &lt;element name="destSTP" type="{http://schemas.ogf.org/nsi/2011/07/connection/types}ServiceTerminationPointType"/>
- *         &lt;element name="stpList" type="{http://schemas.ogf.org/nsi/2011/07/connection/types}StpListType" minOccurs="0"/>
- *         &lt;element name="children" type="{http://schemas.ogf.org/nsi/2011/07/connection/types}ChildListType" minOccurs="0"/>
+ *         &lt;element name="directionality" type="{http://schemas.ogf.org/nsi/2011/10/connection/types}DirectionalityType"/>
+ *         &lt;element name="sourceSTP" type="{http://schemas.ogf.org/nsi/2011/10/connection/types}ServiceTerminationPointType"/>
+ *         &lt;element name="destSTP" type="{http://schemas.ogf.org/nsi/2011/10/connection/types}ServiceTerminationPointType"/>
+ *         &lt;element name="stpList" type="{http://schemas.ogf.org/nsi/2011/10/connection/types}StpListType" minOccurs="0"/>
+ *         &lt;element name="children" type="{http://schemas.ogf.org/nsi/2011/10/connection/types}ChildListType" minOccurs="0"/>
  *       &lt;/sequence>
- *       &lt;attribute name="order" type="{http://www.w3.org/2001/XMLSchema}integer" />
+ *       &lt;attribute name="order" type="{http://www.w3.org/2001/XMLSchema}int" />
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
@@ -77,7 +76,7 @@ public class ComponentPathType {
     protected StpListType stpList;
     protected ChildListType children;
     @XmlAttribute(name = "order")
-    protected BigInteger order;
+    protected Integer order;
 
     /**
      * Gets the value of the directionality property.
@@ -204,10 +203,10 @@ public class ComponentPathType {
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getOrder() {
+    public Integer getOrder() {
         return order;
     }
 
@@ -216,10 +215,10 @@ public class ComponentPathType {
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setOrder(BigInteger value) {
+    public void setOrder(Integer value) {
         this.order = value;
     }
