+++ nsibridge/src/main/java/org/ogf/nsi/schema/connectionTypes/OrderedServiceTerminationPointType.java
@@ -1,7 +1,6 @@
 
 package org.ogf.nsi.schema.connectionTypes;
 
-import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
@@ -33,9 +32,9 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="stpId" type="{http://schemas.ogf.org/nsi/2011/07/connection/types}StpIdType"/>
+ *         &lt;element name="stpId" type="{http://schemas.ogf.org/nsi/2011/10/connection/types}StpIdType"/>
  *       &lt;/sequence>
- *       &lt;attribute name="order" type="{http://www.w3.org/2001/XMLSchema}integer" />
+ *       &lt;attribute name="order" type="{http://www.w3.org/2001/XMLSchema}int" />
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
@@ -52,7 +51,7 @@ public class OrderedServiceTerminationPointType {
     @XmlElement(required = true)
     protected String stpId;
     @XmlAttribute(name = "order")
-    protected BigInteger order;
+    protected Integer order;
 
     /**
      * Gets the value of the stpId property.
@@ -83,10 +82,10 @@ public class OrderedServiceTerminationPointType {
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getOrder() {
+    public Integer getOrder() {
         return order;
     }
 
@@ -95,10 +94,10 @@ public class OrderedServiceTerminationPointType {
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setOrder(BigInteger value) {
+    public void setOrder(Integer value) {
         this.order = value;
     }
